account for changes to main
Need to clean code of tests I was doing, atm slower then previous baseline. Does blockIntensities in DataModellerOutputRecord need to be a 2 dimensional array or 1? Its treated as both in different parts of the code. Current implementation I changed it too is 1 however I'm guessing when multiple blocks are involved will need to be a 2 dimensional array.